# Conflicts:
#	src/map/utils/fishingutils.cpp

+++ src/map/utils/fishingutils.cpp
@@ -404,7 +404,7 @@ namespace fishingutils
     *																		*
     ************************************************************************/
 
-    void FishingAction(CCharEntity* PChar, FISHACTION action, uint16 fightStamina, uint8 special)
+    void FishingAction(CCharEntity* PChar, FISHACTION action, uint16 fightStamina, uint32 special)
     {
         uint16 MessageOffset = GetMessageOffset(PChar->getZone());
